+++ src/brushes/spray_brush.class.ts
@@ -1,6 +1,6 @@
 import { Point } from '../Point';
 import { Group } from '../shapes/Group';
-import { Shadow } from '../shadow.class';
+import { Shadow } from '../Shadow';
 import { Rect } from '../shapes/Rect';
 import { getRandomInt } from '../util/internals';
 import type { Canvas } from '../canvas/canvas_events';
