+++ src/shapes/Object/Object.ts
@@ -4,7 +4,7 @@ import { ALIASING_LIMIT, iMatrix, VERSION } from '../../constants';
 import { ObjectEvents } from '../../EventTypeDefs';
 import { AnimatableObject } from './AnimatableObject';
 import { Point } from '../../Point';
-import { Shadow } from '../../shadow.class';
+import { Shadow } from '../../Shadow';
 import type {
   TClassProperties,
   TDegree,
